refactor: split datasync/downloader/api.go into separate files by API responsibility
- datasync/downloader/api.go → datasync/downloader/api_kaia_downloader.go (KaiaDownloaderAPI)
- datasync/downloader/api.go → datasync/downloader/api_kaia_downloader_sync.go (KaiaDownloaderSyncAPI)
